gazed_object_detectors/modified_detr_default/requirements.txt
lightning==2.0.2
numpy==1.24.3
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
packaging==23.1
pandas==2.0.1
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
PyYAML==6.0
PyYAML==6.0
scipy==1.10.0
torch==2.0.1
torch==2.0.0+cu117
torchmetrics==0.11.4
torchvision==0.15.2
torchvision==0.15.1+cu117
wandb==0.15.2
